refactor(node-ui/pca): #1355 V2 — lift health derivation to a domain module (no layer inversion)
PR #1355 round-1 V2 — coverage.ts (domain) imported healthForSnapshot from
components/Pca/HealthChip.tsx (presentation), a layer inversion. Pure move, no logic change.

- New src/ui/pca/health.ts holds healthForSnapshot + PcaHealthState + PCA_CAP_NEAR_THRESHOLD
  + PCA_EXPIRING_SOON_SECONDS (the consts it uses).
- HealthChip.tsx keeps HEALTH_CHIP_META + the HealthChip component, imports the moved items
  from pca/health.js and RE-EXPORTS them for back-compat.
- Importers repointed to the domain module: coverage.ts, usePcaOverview, usePcaAlerts,
  ConvictionDetailView, components/Pca/index.ts (now re-exports health from pca/health),
  + the two test files (pca-components, pca-mocks). (usePublishEligibility no longer imports
  healthForSnapshot after V1, so nothing to repoint there.)

Behavior-identical; existing tests green. tsc clean; full node-ui suite serial 1671 passed /
38 skipped / 0 failed.
